WIP/API/requirements.txt
fastapi[standard]
redis
pydantic
python-dotenv
pathlib
sqlalchemy
asyncpg